Draft variable density tesseroids implementation
Refactor tesseroid code by moving some functions to a new
_tesseroid_utils.py file.
Create a new _tesseroid_variable_density.py file with density-based
discretization functions and with a gauss_legendre_quadrature function
that works with variable density.
Add new forward modelling functions for variable density tesseroids
only.
Add new example that makes use of variable density tesseroids.
